+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - Support for String literals with escape sequences
+- Added support for multi-line comments (""" comment\n\n comment """)
 
 ### Changed